Aggiunta creazione db e tabella

+++ src/PythonService/models.py
@@ -1,5 +1,5 @@
 from flask_sqlalchemy import SQLAlchemy
-from sqlalchemy import MetaData
+from sqlalchemy import MetaData, create_engine
 from sqlalchemy.ext.declarative import declarative_base
 
 metadata = MetaData()
@@ -9,6 +9,6 @@ db = SQLAlchemy(metadata=metadata)
 class Store(db.Model):
     __tablename__ = "Store"
     productId = db.Column(db.Integer, primary_key=True)
-    availability = db.Column(db.Integer, nullable=False)
+    availability = db.Column(db.Numeric, nullable=False)
     def __repr__(self):
         return '<ProductStore %r>' % self.productId
